Extract socketio to own file
